feat: restructure /components folder
