+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import Recommendation from "@/components/home/Recommendation";
-import ClassList from "@/components/home/ClassList";
+import Recommendation from "@/components/pages/home/Recommendation";
+import ClassList from "@/components/pages/home/ClassList";
 
 export default function HomePage() {
   return (
