+++ src/components/pages/class/carousel/Carousel.tsx
