+++ src/components/pages/class/carousel/CarouselImage.tsx
