+++ src/components/pages/class/checkout/ClassInfo.tsx
