+++ src/components/pages/class/checkout/RefundPolicy.tsx
