+++ src/components/pages/class/reservation/Modal.tsx
