+++ src/components/pages/class/section/SectionContent.tsx
