+++ src/components/pages/class/share/CopyLink.tsx
